refactor(modules): extract last static method, showLoading, from main module
