+++ src/sweetalert2.js
@@ -608,30 +608,6 @@ Object.assign(sweetAlert, staticMethods)
 /**
  * Show spinner instead of Confirm button and disable Cancel button
  */
-sweetAlert.showLoading = sweetAlert.enableLoading = () => {
-  let popup = dom.getPopup()
-  if (!popup) {
-    sweetAlert('')
-  }
-  popup = dom.getPopup()
-  const actions = dom.getActions()
-  const confirmButton = dom.getConfirmButton()
-  const cancelButton = dom.getCancelButton()
-
-  dom.show(actions)
-  dom.show(confirmButton)
-  dom.addClass([popup, actions], swalClasses.loading)
-  confirmButton.disabled = true
-  cancelButton.disabled = true
-
-  popup.setAttribute('data-loading', true)
-  popup.setAttribute('aria-busy', true)
-  popup.focus()
-}
-
-/**
- * Show spinner instead of Confirm button and disable Cancel button
- */
 sweetAlert.hideLoading = sweetAlert.disableLoading = () => {
   if (currentContext) {
     const {params, domCache} = currentContext
